pyproject.toml
[tool.poetry]
name = "pdac-ml"
version = "0.1.0"
description = ""
authors = ["deslobodzian <de.slobodzian@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.2"
pandas = "^2.2.2"
matplotlib = "^3.9.0"
pillow = "^10.3.0"
jupyter = "^1.0.0"
ipykernel = "^6.29.4"
scikit-learn = "^1.5.1"
seaborn = "^0.13.2"
tqdm = "^4.66.5"
grad-cam = "^1.5.3"

[tool.poetry.dependencies.torch]
version = "2.3.1"
source = "pytorch"

[tool.poetry.dependencies.torchvision]
version = "0.18.1"
source = "pytorch"

[tool.poetry.dependencies.torchaudio]
version = "2.3.1"
source = "pytorch"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
